+++ packages/typespec-client-generator-core/src/context.ts
@@ -2,6 +2,7 @@ import {
   createDiagnosticCollector,
   EmitContext,
   emitFile,
+  listServices,
   Model,
   ModelProperty,
   Namespace,
@@ -12,6 +13,7 @@ import {
   Union,
 } from "@typespec/compiler";
 import { HttpOperation } from "@typespec/http";
+import { getVersions } from "@typespec/versioning";
 import { stringify } from "yaml";
 import { defaultDecoratorsAllowList } from "./configs.js";
 import { handleClientExamples } from "./example.js";
@@ -31,6 +33,7 @@ import {
   BrandedSdkEmitterOptionsInterface,
   handleVersioningMutationForGlobalNamespace,
   parseEmitterName,
+  removeVersionsLargerThanExplicitlySpecified,
   TCGCEmitterOptions,
   TspLiteralType,
 } from "./internal-utils.js";
@@ -47,7 +50,8 @@ export function createTCGCContext(program: Program, emitterName?: string): TCGCC
 
     previewStringRegex: /-preview$/,
     disableUsageAccessPropagationToBase: false,
-
+    generateProtocolMethods: true,
+    generateConvenienceMethods: true,
     __referencedTypeCache: new Map<
       Type,
       SdkModelType | SdkEnumType | SdkUnionType | SdkNullableType
@@ -83,11 +87,31 @@ export function createTCGCContext(program: Program, emitterName?: string): TCGCC
       }
       this.__tspTypeToApiVersions.set(type, mergedApiVersions);
     },
+    getPackageVersions(): string[] {
+      if (this.__packageVersions) {
+        return this.__packageVersions;
+      }
+      const service = listServices(program)[0];
+      if (!service) {
+        this.__packageVersions = [];
+        return this.__packageVersions;
+      }
+
+      const versions = getVersions(program, service.type)[1]?.getVersions();
+      if (!versions) {
+        this.__packageVersions = [];
+        return this.__packageVersions;
+      }
+
+      removeVersionsLargerThanExplicitlySpecified(this, versions);
+
+      this.__packageVersions = versions.map((version) => version.value);
+      return this.__packageVersions;
+    },
   };
 }
 
 interface VersioningStrategy {
-  readonly strategy?: "ignore";
   readonly previewStringRegex?: RegExp; // regex to match preview versions
 }
 
@@ -108,12 +132,14 @@ export async function createSdkContext<
   options?: CreateSdkContextOptions,
 ): Promise<SdkContext<TOptions, TServiceOperation>> {
   const diagnostics = createDiagnosticCollector();
-  const generateProtocolMethods = context.options["generate-protocol-methods"] ?? true;
-  const generateConvenienceMethods = context.options["generate-convenience-methods"] ?? true;
   const tcgcContext = createTCGCContext(
     context.program,
     emitterName ?? context.options["emitter-name"],
   );
+  const generateProtocolMethods =
+    context.options["generate-protocol-methods"] ?? tcgcContext.generateProtocolMethods;
+  const generateConvenienceMethods =
+    context.options["generate-convenience-methods"] ?? tcgcContext.generateConvenienceMethods;
   const sdkContext: SdkContext<TOptions, TServiceOperation> = {
     ...tcgcContext,
     emitContext: context,
@@ -122,7 +148,7 @@ export async function createSdkContext<
     generateConvenienceMethods: generateConvenienceMethods,
     examplesDir: context.options["examples-dir"],
     namespaceFlag: context.options["namespace"],
-    apiVersion: options?.versioning?.strategy === "ignore" ? "all" : context.options["api-version"],
+    apiVersion: context.options["api-version"],
     license: context.options["license"],
     decoratorsAllowList: [...defaultDecoratorsAllowList, ...(options?.additionalDecorators ?? [])],
     previewStringRegex: options?.versioning?.previewStringRegex || tcgcContext.previewStringRegex,