+++ packages/typespec-client-generator-core/src/example.ts
@@ -33,7 +33,6 @@ import {
   isSdkFloatKind,
   isSdkIntKind,
 } from "./interfaces.js";
-import { getValidApiVersion } from "./internal-utils.js";
 import { createDiagnostic } from "./lib.js";
 import { getLibraryName } from "./public-utils.js";
 
@@ -168,8 +167,9 @@ export async function handleClientExamples(
 ): Promise<[void, readonly Diagnostic[]]> {
   const diagnostics = createDiagnosticCollector();
 
+  const packageVersions = context.getPackageVersions();
   const examples = diagnostics.pipe(
-    await loadExamples(context, getValidApiVersion(context, client.apiVersions)),
+    await loadExamples(context, packageVersions[packageVersions.length - 1]),
   );
   const clientQueue = [client];
   while (clientQueue.length > 0) {
@@ -406,6 +406,10 @@ function getSdkTypeExample(
 ): [SdkExampleValue | undefined, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
 
+  if (example === null && type.kind !== "nullable" && type.kind !== "unknown") {
+    return diagnostics.wrap(undefined);
+  }
+
   if (isSdkIntKind(type.kind) || isSdkFloatKind(type.kind)) {
     return getSdkBaseTypeExample("number", type as SdkType, example, relativePath);
   } else {
@@ -547,9 +551,6 @@ function getSdkDictionaryExample(
 ): [SdkDictionaryExampleValue | undefined, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   if (typeof example === "object") {
-    if (example === null) {
-      return diagnostics.wrap(undefined);
-    }
     const dictionaryExample: Record<string, SdkExampleValue> = {};
     for (const key of Object.keys(example)) {
       const result = diagnostics.pipe(
@@ -577,9 +578,6 @@ function getSdkModelExample(
 ): [SdkModelExampleValue | undefined, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   if (typeof example === "object") {
-    if (example === null) {
-      return diagnostics.wrap(undefined);
-    }
     // handle discriminated model
     if (type.discriminatorProperty) {
       if (