+++ packages/typespec-client-generator-core/src/http.ts
@@ -6,10 +6,8 @@ import {
   Type,
   compilerAssert,
   createDiagnosticCollector,
-  getDoc,
   getEncode,
   getSummary,
-  ignoreDiagnostics,
   isErrorModel,
 } from "@typespec/compiler";
 import { $ } from "@typespec/compiler/experimental/typekit";
@@ -56,6 +54,7 @@ import {
 import {
   findRootSourceProperty,
   getAvailableApiVersions,
+  getClientDoc,
   getHttpBodySpreadModel,
   getHttpOperationResponseHeaders,
   getLocationOfOperation,
@@ -89,6 +88,7 @@ export function getSdkHttpOperation(
   httpOperation: HttpOperation,
   methodParameters: SdkMethodParameter[],
 ): [SdkHttpOperation, readonly Diagnostic[]] {
+  const tk = $(context.program);
   const diagnostics = createDiagnosticCollector();
   const { responses, exceptions } = diagnostics.pipe(
     getSdkHttpResponseAndExceptions(context, httpOperation),
@@ -97,11 +97,11 @@ export function getSdkHttpOperation(
     // we make sure valid responses and 404 responses are booleans
     for (const response of responses) {
       // all valid responses will return boolean
-      response.type = getSdkConstant(context, $.literal.createBoolean(true));
+      response.type = getSdkConstant(context, tk.literal.createBoolean(true));
     }
     const fourOFourResponse = exceptions.find((e) => e.statusCodes === 404);
     if (fourOFourResponse) {
-      fourOFourResponse.type = getSdkConstant(context, $.literal.createBoolean(false));
+      fourOFourResponse.type = getSdkConstant(context, tk.literal.createBoolean(false));
       // move from exception to valid response with status code 404
       responses.push({
         ...fourOFourResponse,
@@ -114,7 +114,7 @@ export function getSdkHttpOperation(
       responses.push({
         kind: "http",
         statusCodes: 404,
-        type: getSdkConstant(context, $.literal.createBoolean(false)),
+        type: getSdkConstant(context, tk.literal.createBoolean(false)),
         apiVersions: getAvailableApiVersions(
           context,
           httpOperation.operation,
@@ -215,6 +215,7 @@ function getSdkHttpParameters(
     } else if (!isNeverOrVoidType(tspBody.type)) {
       const spread = isHttpBodySpread(tspBody);
       let type: SdkType;
+      let optional = false;
       if (spread) {
         type = diagnostics.pipe(
           getClientTypeWithDiagnostics(
@@ -227,6 +228,7 @@ function getSdkHttpParameters(
         type = diagnostics.pipe(
           getClientTypeWithDiagnostics(context, tspBody.type, httpOperation.operation),
         );
+        optional = tspBody.property?.optional ?? false;
       }
       const name = camelCase((type as { name: string }).name ?? "body");
       retval.bodyParam = {
@@ -234,7 +236,7 @@ function getSdkHttpParameters(
         name,
         isGeneratedName: true,
         serializedName: "",
-        doc: getDoc(context.program, tspBody.type),
+        doc: getClientDoc(context, tspBody.type),
         summary: getSummary(context.program, tspBody.type),
         onClient: false,
         contentTypes: [],
@@ -242,7 +244,7 @@ function getSdkHttpParameters(
         isApiVersionParam: false,
         apiVersions: getAvailableApiVersions(context, tspBody.type, httpOperation.operation),
         type,
-        optional: false,
+        optional,
         correspondingMethodParams,
         crossLanguageDefinitionId: `${getCrossLanguageDefinitionId(context, httpOperation.operation)}.body`,
         decorators: diagnostics.pipe(getTypeDecorators(context, tspBody.type)),
@@ -426,16 +428,10 @@ export function getSdkHttpParameter(
       // url type need allow reserved
       allowReserved:
         (httpParam as HttpOperationPathParameter)?.allowReserved ??
-        ignoreDiagnostics(
-          program.checker.isTypeAssignableTo(
-            param.type,
-            program.checker.getStdType("url"),
-            param.type,
-          ),
-        ),
+        $(program).type.isAssignableTo(param.type, $(program).builtin.url, param.type),
       serializedName: getPathParamName(program, param) ?? base.name,
       correspondingMethodParams,
-      optional: false,
+      optional: param.optional,
     });
   }
   if (isCookieParam(context.program, param) || location === "cookie") {
@@ -804,7 +800,7 @@ function getFormatFromExplodeOrEncode(
   explode?: boolean,
 ): [CollectionFormat | undefined, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
-  if ($.array.is(type.type)) {
+  if ($(context.program).array.is(type.type)) {
     if (explode) {
       return diagnostics.wrap("multi");
     }