+++ packages/typespec-client-generator-core/src/package.ts
@@ -1,22 +1,23 @@
-import { createDiagnosticCollector, Diagnostic } from "@typespec/compiler";
-import { resolveVersions } from "@typespec/versioning";
+import { createDiagnosticCollector, Diagnostic, DiagnosticCollector } from "@typespec/compiler";
 import { createSdkClientType } from "./clients.js";
 import { listClients, listOperationGroups } from "./decorators.js";
 import {
-  SdkClient,
+  SdkClientType,
   SdkEnumType,
   SdkModelType,
   SdkNamespace,
   SdkNullableType,
-  SdkOperationGroup,
   SdkPackage,
   SdkServiceOperation,
   SdkUnionType,
   TCGCContext,
 } from "./interfaces.js";
-import { filterApiVersionsWithDecorators } from "./internal-utils.js";
+import {
+  filterApiVersionsWithDecorators,
+  hasExplicitClientOrOperationGroup,
+} from "./internal-utils.js";
 import { getLicenseInfo } from "./license.js";
-import { getCrossLanguagePackageId, getDefaultApiVersion } from "./public-utils.js";
+import { getCrossLanguagePackageId } from "./public-utils.js";
 import { getAllReferencedTypes, handleAllTypes } from "./types.js";
 
 export function createSdkPackage<TServiceOperation extends SdkServiceOperation>(
@@ -27,8 +28,9 @@ export function createSdkPackage<TServiceOperation extends SdkServiceOperation>(
   diagnostics.pipe(handleAllTypes(context));
   const crossLanguagePackageId = diagnostics.pipe(getCrossLanguagePackageId(context));
   const allReferencedTypes = getAllReferencedTypes(context);
+  const versions = context.getPackageVersions();
   const sdkPackage: SdkPackage<TServiceOperation> = {
-    clients: listClients(context).map((c) => diagnostics.pipe(createSdkClientType(context, c))),
+    clients: filterClients(context, diagnostics),
     models: allReferencedTypes.filter((x): x is SdkModelType => x.kind === "model"),
     enums: allReferencedTypes.filter((x): x is SdkEnumType => x.kind === "enum"),
     unions: allReferencedTypes.filter(
@@ -37,9 +39,26 @@ export function createSdkPackage<TServiceOperation extends SdkServiceOperation>(
     crossLanguagePackageId,
     namespaces: [],
     licenseInfo: getLicenseInfo(context),
+    metadata: {
+      apiVersion: context.apiVersion === "all" ? "all" : versions[versions.length - 1],
+    },
   };
   organizeNamespaces(sdkPackage);
   return diagnostics.wrap(sdkPackage);
+}
+
+function filterClients<TServiceOperation extends SdkServiceOperation>(
+  context: TCGCContext,
+  diagnostics: DiagnosticCollector,
+): SdkClientType<TServiceOperation>[] {
+  const allClients: SdkClientType<TServiceOperation>[] = listClients(context).map((c) =>
+    diagnostics.pipe(createSdkClientType(context, c)),
+  );
+  if (hasExplicitClientOrOperationGroup(context)) {
+    return allClients;
+  } else {
+    return allClients.filter((c) => c.methods.length > 0 || (c.children && c.children.length > 0));
+  }
 }
 
 function organizeNamespaces<TServiceOperation extends SdkServiceOperation>(
@@ -97,41 +116,27 @@ function getSdkNamespace<TServiceOperation extends SdkServiceOperation>(
 
 function populateApiVersionInformation(context: TCGCContext): void {
   for (const client of listClients(context)) {
-    let clientApiVersions = resolveVersions(context.program, client.service)
-      .filter((x) => x.rootVersion)
-      .map((x) => x.rootVersion!.value);
     context.setApiVersionsForType(
       client.type,
-      filterApiVersionsWithDecorators(context, client.type, clientApiVersions),
+      filterApiVersionsWithDecorators(context, client.type, context.getPackageVersions()),
     );
 
+    const clientApiVersions = context.getApiVersionsForType(client.type);
     context.__clientToApiVersionClientDefaultValue.set(
       client.type,
-      getClientDefaultApiVersion(context, client),
+      clientApiVersions[clientApiVersions.length - 1],
     );
     for (const sc of listOperationGroups(context, client, true)) {
-      clientApiVersions = resolveVersions(context.program, sc.service)
-        .filter((x) => x.rootVersion)
-        .map((x) => x.rootVersion!.value);
       context.setApiVersionsForType(
         sc.type,
-        filterApiVersionsWithDecorators(context, sc.type, clientApiVersions),
+        filterApiVersionsWithDecorators(context, sc.type, context.getPackageVersions()),
       );
 
+      const clientApiVersions = context.getApiVersionsForType(sc.type);
       context.__clientToApiVersionClientDefaultValue.set(
         sc.type,
-        getClientDefaultApiVersion(context, sc),
+        clientApiVersions[clientApiVersions.length - 1],
       );
     }
   }
-}
-
-function getClientDefaultApiVersion(
-  context: TCGCContext,
-  client: SdkClient | SdkOperationGroup,
-): string | undefined {
-  if (context.apiVersion && !["latest", "all"].includes(context.apiVersion)) {
-    return context.apiVersion;
-  }
-  return getDefaultApiVersion(context, client.service)?.value;
 }