+++ packages/typespec-azure-rulesets/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - @azure-tools/typespec-azure-rulesets
 
+## 0.55.0
+
+No changes, version bump only.
+
 ## 0.54.0
 
 ### Features